Added Test for devices module
