+++ pybrid/data.py
@@ -7,11 +7,6 @@ def get_status(auth):
     data = pybrid.conn.get_data(auth, base_url, data_url)
     return data
 
-def get_device_info(auth):
-    base_url = "http://" + str(auth.ipaddress)
-    data_url = "/info"
-    data = pybrid.conn.get_data(auth, base_url, data_url)
-    return data
 
 def get_history(auth):
     base_url = "http://" + str(auth.ipaddress)
